@@ -1,8 +1,8 @@
 plugins {
-    id 'org.jetbrains.intellij' version '1.16.1'
+    id 'org.jetbrains.intellij' version '1.17.4'
     id 'org.jetbrains.changelog' version '2.2.0'
     id 'com.github.ldenisey.setversions' version '1.0.4'
-    id "com.github.node-gradle.node" version "7.0.1"
+    id "com.github.node-gradle.node" version "7.0.2"
 }
 
 apply plugin: 'org.jetbrains.changelog'
@@ -10,8 +10,8 @@ apply plugin: 'org.jetbrains.changelog'
 group = 'com.javampire'
 version = '2.4.1-SNAPSHOT'
 
-sourceCompatibility = 1.11
-targetCompatibility = 1.11
+sourceCompatibility = 17
+targetCompatibility = 17
 sourceSets.main.java.srcDirs = ['src/main/java', 'src/main/gen']
 sourceSets.main.resources.srcDirs = ['src/main/resources', 'src/main/generated-resources']
 
@@ -37,7 +37,8 @@ intellij {
     updateSinceUntilBuild.set(false)
 
     patchPluginXml {
-        sinceBuild.set('212.5080')
+//        sinceBuild.set('212.5080')
+        sinceBuild.set('232')
         untilBuild.set('')
 
         changeNotes.set(provider { changelog.getLatest().toHTML() })
